+++ pom.xml
@@ -8,11 +8,11 @@
 		<version>3.4.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.example</groupId>
-	<artifactId>muslce-backend-app</artifactId>
+	<groupId>com.github.ten-revengineer</groupId>
+	<artifactId>muscle-backend-app</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>springboot-hello</name>
-	<description>Demo project for Spring Boot</description>
+	<name>muscle-backend-app</name>
+	<description>Spring Boot backend for muscle tracking</description>
 	<url/>
 	<licenses>
 		<license/>
