+++ src/main/java/com/github/ten_revengineer/muscle/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.springboot_hello.repository;
+package com.github.ten_revengineer.muscle.repository;
 
-import com.example.springboot_hello.model.User;
+import com.github.ten_revengineer.muscle.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
