+++ src/test/java/com/github/ten_revengineer/muscle/MuscleBackendApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.springboot_hello;
+package com.github.ten_revengineer.muscle;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SpringbootHelloApplicationTests {
+class MuscleBackendApplicationTests {
 
 	@Test
 	void contextLoads() {
